extract applyRoles, fix github build
