+++ packages/github/package.json
@@ -19,13 +19,14 @@
   "dependencies": {
     "@hey-amplify/constants": "workspace:*",
     "@hey-amplify/prisma-client": "workspace:*",
-    "@hey-amplify/discord": "workspace:*",
     "@octokit/auth-app": "^4.0.9",
     "@octokit/graphql": "^5.0.5",
     "@octokit/rest": "^19.0.7"
   },
   "devDependencies": {
+    "@octokit/types": "^9.0.0",
     "tsup": "^6.5.0",
-    "typescript": "^4.9.5"
+    "typescript": "^4.9.5",
+    "vitest": "^0.28.4"
   }
 }
